refactor(reef): use wax component barrel imports
