+++ apps/reef/app/views/traces/http-span-detail.tsx
@@ -1,8 +1,7 @@
 import { useCallback, useEffect, useState } from 'react'
 import classNames from 'classnames'
 
-import { ScrollArea } from '@/wax/components'
-import * as Button from '@/wax/components/button'
+import { Button, ScrollArea } from '@/wax/components'
 import { KeyboardShortcut } from '@/wax/components/keyboard-shortcut'
 import { Typography } from '@/wax/components/typography'
 import type { TraceSpan } from '@/generated/coral/v1/traces_pb'
