+++ apps/reef/app/wax/components/index.ts
@@ -7,5 +7,6 @@ export * as KeyboardShortcut from './keyboard-shortcut'
 export * as List from './list'
 export * as Menu from './menu'
 export * as ScrollArea from './scroll-area'
+export * as Tabs from './tabs'
 export { Table } from './table'
 export { Typography, TypographyVariant } from './typography'
