+++ apps/reef/app/wax/components/list/list.stories.tsx
@@ -3,9 +3,8 @@ import type { Meta, StoryObj } from '@storybook/react-vite'
 import { fn } from 'storybook/test'
 
 import { utils } from '@/styles/utils'
+import { List } from '@/wax/components'
 import { theme } from '@/wax/theme/theme.css'
-
-import * as List from './index'
 
 const meta: Meta<typeof List.Container> = {
   component: List.Container,
